Remove whitespace
While I was trying to understand how this bot works, I was being constantly distracted by an unknown source of annoyance. I couldn't figure out if it was the lack of brackets, the weird comment syntax or the fact it was written in python.

After a few sleepless nights, I woke up in cold sweat at 3:35 in the morning and went to check the source code again. And I was right. It was revealed to me in a dream, that there was an occurrence of a mortal sin on line 9. 

To spare others of such fate, I hereby fix this error and provide you this PR.